// src/app/interfaces/estudiante.ts
export interface Estudiante {
    activo: boolean;
    apellidos: string;
    cv: string;
    email: string;
    empresa: string;
    estudios: string;
    experiencia: string;
    fechaAlta: string;       
    fechaNacimiento: string;
    fotoPerfil: string;
    nombre: string;
    rol: string;
    telefono: string;
    username: string;
}