// src/app/interfaces/registro-estudiante-dto.ts
export interface RegistroEstudianteDto {
  nombre: string;
  apellidos: string;
  email: string;
  estudios: string;
  experiencia: string;
  cv: string;
  fotoPerfil: string;
  telefono: string;
  username: string;
  password: string;
  fechaNacimiento: Date;
  calle: string;
  pais: string;
  codigoPostal: string;
  provincia: string;
  poblacion: string;
}